# contracts/openapi/public-api.yaml
openapi: 3.1.0
info:
  title: Hootly Public API
  version: 0.6.0
  description: >
    Contrat public Hootly. Authentification (Sprint 02), profil et marques
    (Sprint 03), publications, médias, planification et calendrier (Sprint 04),
    comptes sociaux et OAuth Meta (Sprint 06/07), commentaires et réponses
    contrôlées (Sprint 08), analyse de sentiment (Sprint 09), propositions de
    réponse et hashtags (Sprint 10), notifications et Firebase Cloud
    Messaging (Sprint 11).
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/notifications:
    get:
      summary: Lister les notifications de l'utilisateur connecté (Sprint 11)
      description: >
        Portée par l'utilisateur, jamais par la marque directement : une
        notification est créée une fois par destinataire (voir Jour 3), donc
        l'appartenance à `userId` EST le contrôle d'accès. `filter` reprend
        exactement la sémantique de `notificationsApi.list()` côté mobile
        (`all`, `unread`, `priority` = priorité haute, `errors` = échecs de
        publication/synchronisation/token).
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: filter, in: query, schema: { type: string, enum: [all, unread, priority, errors], default: all } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
  /api/v1/notifications/unread-count:
    get:
      summary: Compteur de notifications non lues (Sprint 11)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
  /api/v1/notifications/{notificationId}/read:
    parameters:
      - { name: notificationId, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      summary: Marquer une notification comme lue (Sprint 11)
      description: Idempotent — relire une notification déjà lue ne change pas son horodatage de lecture.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/notifications/read-all:
    post:
      summary: Marquer toutes les notifications non lues comme lues (Sprint 11)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/AuthenticationRequired' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta:
          type: object
          required: [requestId]
          properties:
            requestId: { type: string }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: array, items: { type: object } }
  responses:
...
    Success:
      description: RÃ©ponse rÃ©ussie sous enveloppe Hootly
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
...
    ValidationFailed:
      description: Entrée invalide
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    AuthenticationRequired:
      description: Authentification invalide ou absente
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
    NotFound:
      description: Ressource introuvable ou non accessible
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...